pyproject.toml
[project]
name = "ptu-ai-powered-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.4",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "groq>=0.31.0",
    "markdown>=3.8.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
]
